requirements.txt
diff-match-patch
wxpython
pycryptodome
